Changed some asserts and we are all done

+++ project2D/Setting.cpp
@@ -51,9 +51,9 @@ void Setting::update(float deltaTime, StateManager * SM)
 
 	}
 
-	if (leftPaddle->posY + leftPaddle->tall / 2 > app->getWindowHeight())
+	if (leftPaddle->posY + leftPaddle->tall / 2 > (float)app->getWindowHeight())
 	{
-		leftPaddle->posY = app->getWindowHeight() - leftPaddle->tall / 2;
+		leftPaddle->posY = (float)app->getWindowHeight() - leftPaddle->tall / 2;
 
 	}
 
@@ -62,40 +62,40 @@ void Setting::update(float deltaTime, StateManager * SM)
 		rightPaddle->posY = rightPaddle->tall / 2;
 	}
 
-	if (rightPaddle->posY + rightPaddle->tall / 2 > app->getWindowHeight())
+	if (rightPaddle->posY + rightPaddle->tall / 2 > (float)app->getWindowHeight())
 	{
-		rightPaddle->posY = app->getWindowHeight() - rightPaddle->tall / 2;
+		rightPaddle->posY = (float)app->getWindowHeight() - rightPaddle->tall / 2;
 	}
 
 
 	if (ball->posY + ball->tall / 2 > app->getWindowHeight())
 	{
-		ball->Yvelocity = ball->Yvelocity * -1;
+		ball->Yvelocity = ball->Yvelocity * -1.0f;
 	}
 
 	if (ball->posY - ball->tall / 2 < 0)
 	{
-		ball->Yvelocity = ball->Yvelocity * -1;
+		ball->Yvelocity = ball->Yvelocity * -1.0f;
 	}
 
 	if (checkCollision(ball->posX, ball->posY, ball->wide, ball->tall, rightPaddle->posX, rightPaddle->posY, rightPaddle->wide, rightPaddle->tall))
 	{
-		ball->Xvelocity = ball->Xvelocity * -1.5;
+		ball->Xvelocity = ball->Xvelocity * -1.5f;
 
 	}
 
 	if (checkCollision(ball->posX, ball->posY, ball->wide, ball->tall, leftPaddle->posX, leftPaddle->posY, leftPaddle->wide, leftPaddle->tall))
 	{
-		ball->Xvelocity = ball->Xvelocity * -1.5;
+		ball->Xvelocity = ball->Xvelocity * -1.5f;
 
 	}
 
-	if (ball->posX > app->getWindowWidth() || ball->posX < 0)
+	if (ball->posX > (float)app->getWindowWidth() || ball->posX < 0)
 	{
-		ball->posX = app->getWindowWidth() / 2;
-		ball->posY = app->getWindowHeight() / 2;
-		ball->Yvelocity = 200;
-		ball->Xvelocity = 200;
+		ball->posX = (float)app->getWindowWidth() / 2;
+		ball->posY = (float)app->getWindowHeight() / 2;
+		ball->Yvelocity = 200.0f;
+		ball->Xvelocity = 200.0f;
 	}
 
 	ball->posX += ball->Xvelocity * deltaTime;
@@ -124,19 +124,19 @@ Setting::~Setting()
 }
 
 
-bool Setting::checkCollision(int x, int y, int oWidth, int oHeight, int xTwo, int yTwo, int oTwoWidth, int oTwoHeight)
+bool Setting::checkCollision(float x, float y, float oWidth, float oHeight, float xTwo, float yTwo, float oTwoWidth, float oTwoHeight)
 {
 
-	int x1Min = x - oWidth / 2;
-	int x1Max = x + oWidth / 2;
-	int y1Max = y + oHeight / 2;
-	int y1Min = y = oHeight / 2;
+	float x1Min = x - oWidth / 2;
+	float x1Max = x + oWidth / 2;
+	float y1Max = y + oHeight / 2;
+	float y1Min = y = oHeight / 2;
 
 	// AABB 2
-	int x2Min = xTwo - oTwoWidth / 2;
-	int x2Max = xTwo + oTwoWidth / 2;
-	int y2Max = yTwo + oTwoHeight / 2;
-	int y2Min = yTwo = oTwoHeight / 2;
+	float x2Min = xTwo - oTwoWidth / 2;
+	float x2Max = xTwo + oTwoWidth / 2;
+	float y2Max = yTwo + oTwoHeight / 2;
+	float y2Min = yTwo = oTwoHeight / 2;
 
 	// Collision tests
 	if (x1Max< x2Min || x1Min > x2Max) return false;
